% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.doc.images-package.R
\docType{package}
\name{gt.doc.images-package}
\alias{gt.doc.images}
\alias{gt.doc.images-package}
\title{gt.doc.images: Save Help File Images}
\description{
Exports tools for creating help file documents that contain images of 'gtsummary', 'gt', and 'flextable' images.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.danieldsjoberg.com/gt.doc.images/}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com}

}
\keyword{internal}
