% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_help_file_images.R
\name{save_help_file_images}
\alias{save_help_file_images}
\title{Save Rd File Images}
\usage{
save_help_file_images(
  path = here::here(),
  pkg = basename(path),
  rd_files = NULL,
  delete_existing_pngs = NULL,
  run.dontrun = TRUE,
  run.donttest = TRUE
)
}
\arguments{
\item{path}{path to package directory. Default is \code{here::here()}}

\item{pkg}{Name of package. Default is the folder name in \verb{path=}}

\item{rd_files}{Character vector of Rd file names to search. Default is all
Rd files. Include the \code{.Rd} extension in the file names.}

\item{delete_existing_pngs}{Logical indicating whether to delete all
png files in the \verb{~/man/figures/} folder. Default is to ask interactively
when \code{rd_files = NULL}. Otherwise, \code{FALSE}}

\item{run.dontrun}{logical indicating that \verb{\\dontrun}
    should be ignored.}

\item{run.donttest}{logical indicating that \verb{\\donttest}
    should be ignored.}
}
\description{
Only objects of class gtsummary, gt, or flextable will be saved.
Object names must end in \verb{_ex} or \verb{_ex[:digit:]}
}
