% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_help_file_images.R
\name{shrink_help_file_images}
\alias{shrink_help_file_images}
\title{Shrink Rd File Images}
\usage{
shrink_help_file_images(
  path = here::here(),
  pkg = basename(path),
  image_files = NULL,
  path.output = file.path(path, "man", "figures")
)
}
\arguments{
\item{path}{path to package directory. Default is \code{here::here()}}

\item{pkg}{Name of package. Default is the folder name in \verb{path=}}

\item{image_files}{Character vector of .png file names to shrink. Default is all
png files in the \verb{~/man/figures} folder that end with \verb{_ex[digits].png}
or \verb{_ex.png} that do not begin with \verb{README-}.}

\item{path.output}{directory where images will be saved. Default is
\code{file.path(path, "man", "figures")}}
}
\description{
Shrink Rd File Images
}
