% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_help_file_images.R
\name{save_help_file_images}
\alias{save_help_file_images}
\title{Save Rd File Images}
\usage{
save_help_file_images(delete_existing_pngs = FALSE)
}
\arguments{
\item{delete_existing_pngs}{Logical}
}
\description{
Only objects of class gtsummary, gt, or flextable will be saved.
Object names must end in \verb{_ex} or \verb{_ex[:digit:]}
}
